pyproject.toml
[project]
name = "sarvam-api"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "ipykernel>=6.29.5",
    "python-dotenv>=1.1.1",
    "sarvamai>=0.1.10",
    "sounddevice>=0.5.2",
    "soundfile>=0.13.1",
    "streamlit>=1.46.1",

]
